rename central_scanner_app to app for consistency across apps
